most of aws china support, refactor ec2
